Extract reusable Screen component from screens
